% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mpg.R
\name{predict_mpg}
\alias{predict_mpg}
\title{Predict the mpg of a car from the weight and number of cylinders}
\usage{
predict_mpg(cyl, wt)
}
\arguments{
\item{cyl}{Number of cylinders, usually 4, 6 or 8}

\item{wt}{Weight of the vehicle, in 1000 pounds}
}
\value{
Predicted mpg
}
\description{
Predict the mpg of a car from the weight and number of cylinders
}
\examples{
predict_mpg(cyl = 6, wt = 3.6)
}
